Remove the CommandBus interface, restructure handler
