+++ examples/simple/simple_test.go
@@ -22,7 +22,7 @@ import (
 	"time"
 
 	eh "github.com/looplab/eventhorizon"
-	commandbus "github.com/looplab/eventhorizon/commandbus/local"
+	"github.com/looplab/eventhorizon/commandhandler/bus"
 	eventbus "github.com/looplab/eventhorizon/eventbus/local"
 	eventstore "github.com/looplab/eventhorizon/eventstore/memory"
 	eventpublisher "github.com/looplab/eventhorizon/publisher/local"
@@ -41,7 +41,7 @@ func Example() {
 	eventBus.SetPublisher(eventPublisher)
 
 	// Create the command bus.
-	commandBus := commandbus.NewCommandBus()
+	commandBus := bus.NewCommandHandler()
 
 	// Create the read repositories.
 	invitationRepo := repo.NewRepo()
